+++ src/main/java/com/plivo/api/models/base/ListResponse.java
@@ -17,6 +17,7 @@ public class ListResponse<T> extends BaseResponse {
   private List<T> profiles;
   private List<T> campaigns;
   private List<T> sessions;
+  private Response response;
 
   /**
    * @return The pagination-related metadata for this list response.
@@ -45,5 +46,8 @@ public class ListResponse<T> extends BaseResponse {
   }
   public List<T> getSessions() {
     return sessions;
+  }
+  public Response getResponse() {
+    return response;
   }
 }