+++ src/main/java/com/plivo/api/models/base/Response.java
@@ -1,0 +1,17 @@
+package com.plivo.api.models.base;
+
+import java.util.List;
+
+public class Response {
+  private Meta meta;
+  private Object objects;
+
+  public Meta getMeta() {
+    return meta;
+  }
+
+  public Object getObjects() {
+    return objects;
+  }
+
+}